+++ .eslintrc.js
@@ -152,6 +152,7 @@ module.exports = {
         'no-underscore-dangle': 'off',
         'no-unneeded-ternary': 'off',
         'no-useless-escape': 'off',
+        'no-void': 'off',
         'no-warning-comments': 'off',
         'object-property-newline': 'off',
         'object-shorthand': [