+++ src/files/BrsFile.Class.spec.ts
@@ -829,6 +829,50 @@ describe('BrsFile BrighterScript classes', () => {
         ]);
     });
 
+    it('detects direct circular extends', () => {
+        //direct
+        program.setFile('source/Direct.bs', `
+            class Parent extends Child
+            end class
+
+            class Child extends Parent
+            end class
+        `);
+        program.validate();
+        expect(
+            program.getDiagnostics().map(x => x.message).sort()
+        ).to.eql([
+            DiagnosticMessages.circularReferenceDetected(['Child', 'Parent', 'Child'], 'source').message,
+            DiagnosticMessages.circularReferenceDetected(['Parent', 'Child', 'Parent'], 'source').message
+        ]);
+    });
+
+    it('detects indirect circular extends', () => {
+        //direct
+        program.setFile('source/Indirect.bs', `
+            class Parent extends Grandchild
+            end class
+
+            class Child extends Parent
+            end class
+
+            class Grandchild extends Child
+            end class
+        `);
+        program.validate();
+        expect(
+            program.getDiagnostics().map(x => x.message).sort()
+        ).to.eql([
+            DiagnosticMessages.circularReferenceDetected(['Child', 'Parent', 'Grandchild', 'Child'], 'source').message,
+            DiagnosticMessages.circularReferenceDetected(['Grandchild', 'Child', 'Parent', 'Grandchild'], 'source').message,
+            DiagnosticMessages.circularReferenceDetected(['Parent', 'Grandchild', 'Child', 'Parent'], 'source').message
+        ]);
+    });
+
+    it('does not cause infinite loop', () => {
+
+    });
+
     it('detects duplicate member names', () => {
         program.setFile('source/main.bs', `
             class Animal